requirements.txt
flask
werkzeug
torch
open_clip_torch
pandas
Pillow
scikit-learn
